+++ services/src/main/java/org/keycloak/protocol/oidc/grants/device/endpoints/DeviceEndpoint.java
@@ -142,10 +142,14 @@ public class DeviceEndpoint extends AuthorizationEndpointBase implements RealmRe
         AuthorizationEndpointChecker checker = new AuthorizationEndpointChecker()
                 .event(event)
                 .client(client)
-                .request(request);
+                .realm(realm)
+                .request(request)
+                .session(session);
 
         try {
             checker.checkPKCEParams(false);
+            checker.checkValidScope();
+            checker.checkValidResource();
         } catch (AuthorizationCheckException ex) {
             throw new ErrorResponseException(ex.getError(), ex.getErrorDescription(), Response.Status.BAD_REQUEST);
         }